// src/app/shared/util-model/enumeration/project-authority.enum.ts
export enum ProjectAuthorityEnum {
    REGISTRY_PROJECT_R = 'REGISTRY_PROJECT_R',
    REGISTRY_PROJECT_U = 'REGISTRY_PROJECT_U',
    REGISTRY_PROJECT_D = 'REGISTRY_PROJECT_D',
    REGISTRY_PROJECT_PROFILE_C = 'REGISTRY_PROJECT_PROFILE_C',
    REGISTRY_PROJECT_PROFILE_R = 'REGISTRY_PROJECT_PROFILE_R',
    REGISTRY_PROJECT_PROFILE_U = 'REGISTRY_PROJECT_PROFILE_U',
    REGISTRY_PROJECT_PROFILE_D = 'REGISTRY_PROJECT_PROFILE_D',
    REGISTRY_PROJECT_PROFILE_METADATA_R = 'REGISTRY_PROJECT_PROFILE_METADATA_R',
    REGISTRY_PROJECT_PARTICIPANT_C = 'REGISTRY_PROJECT_PARTICIPANT_C',
    REGISTRY_PROJECT_PARTICIPANT_R = 'REGISTRY_PROJECT_PARTICIPANT_R',
    REGISTRY_PROJECT_PARTICIPANT_HISTORY_R = 'REGISTRY_PROJECT_PARTICIPANT_HISTORY_R',
    REGISTRY_PROJECT_PARTICIPANT_U = 'REGISTRY_PROJECT_PARTICIPANT_U',
    REGISTRY_PROJECT_PARTICIPANT_D = 'REGISTRY_PROJECT_PARTICIPANT_D',
    REGISTRY_PROJECT_PARTICIPANT_METADATA_R = 'REGISTRY_PROJECT_PARTICIPANT_METADATA_R',
    REGISTRY_PROJECT_GROUP_C = 'REGISTRY_PROJECT_GROUP_C',
    REGISTRY_PROJECT_GROUP_R = 'REGISTRY_PROJECT_GROUP_R',
    REGISTRY_PROJECT_GROUP_U = 'REGISTRY_PROJECT_GROUP_U',
    REGISTRY_PROJECT_GROUP_D = 'REGISTRY_PROJECT_GROUP_D',
    REGISTRY_PROJECT_GROUP_METADATA_R = 'REGISTRY_PROJECT_GROUP_METADATA_R',
    REGISTRY_PROJECT_MOVEMENT_C = 'REGISTRY_PROJECT_MOVEMENT_C',
    REGISTRY_PROJECT_MOVEMENT_R = 'REGISTRY_PROJECT_MOVEMENT_R',
    REGISTRY_PROJECT_MOVEMENT_U = 'REGISTRY_PROJECT_MOVEMENT_U',
    REGISTRY_PROJECT_MOVEMENT_D = 'REGISTRY_PROJECT_MOVEMENT_D',
    REGISTRY_PROJECT_MOVEMENT_METADATA_R = 'REGISTRY_PROJECT_MOVEMENT_METADATA_R',
    REGISTRY_PROJECT_VEHICLE_C = 'REGISTRY_PROJECT_VEHICLE_C',
    REGISTRY_PROJECT_VEHICLE_R = 'REGISTRY_PROJECT_VEHICLE_R',
    REGISTRY_PROJECT_VEHICLE_HISTORY_R = 'REGISTRY_PROJECT_VEHICLE_HISTORY_R',
    REGISTRY_PROJECT_VEHICLE_U = 'REGISTRY_PROJECT_VEHICLE_U',
    REGISTRY_PROJECT_VEHICLE_D = 'REGISTRY_PROJECT_VEHICLE_D',
    REGISTRY_PROJECT_ACTIVITY_C = 'REGISTRY_PROJECT_ACTIVITY_C',
    REGISTRY_PROJECT_ACTIVITY_R = 'REGISTRY_PROJECT_ACTIVITY_R',
    REGISTRY_PROJECT_ACTIVITY_HISTORY_R = 'REGISTRY_PROJECT_ACTIVITY_HISTORY_R',
    REGISTRY_PROJECT_ACTIVITY_U = 'REGISTRY_PROJECT_ACTIVITY_U',
    REGISTRY_PROJECT_ACTIVITY_D = 'REGISTRY_PROJECT_ACTIVITY_D',
}